TV: Move PlaylistDetailViewModel to its own file
